refactor caching into separate files. fix missing spark .cache() return assign in cacheToDbfs()
