+++ pyproject.toml
@@ -13,9 +13,6 @@ dependencies = [
     "pydantic>=2.10.4",
     "pydantic-settings>=2.7.0",
     "tqdm>=4.67.1",
-    "mypy>=1.14.1",
-    "pandas-stubs>=2.2.3.241126",
-    "types-tqdm>=4.67.0.20241221",
     "pyarrow>=19.0.1",
     "typing-extensions==4.12.2",
 ]
@@ -78,12 +75,14 @@ module = ["pyspark.*", "py4j.*"]
 ignore_missing_imports = true
 
 [tool.ruff]
+exclude = [".venv", "build"]
+
+[tool.ruff.lint]
 # Ignore E402 (module level import not at top of file) in test files
 # Ignore F401 (unused imports) in test files
 ignore = ["E402", "E701"]
-exclude = [".venv", "build"]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["F401"]
 "tests/notebooks/*" = ["F821"]
